+++ dexs/thirdfy.ts
@@ -1,6 +1,6 @@
 import type { FetchOptions, SimpleAdapter } from "../adapters/types";
 import { CHAIN } from "../helpers/chains";
-import { request, gql } from "graphql-request";
+import { request } from "graphql-request";
 
 const SUBGRAPH_URL = "https://api.studio.thegraph.com/query/47039/thirdfy-base/version/latest";
 
@@ -17,6 +17,7 @@ interface PoolDayData {
 
 const fetch = async (options: FetchOptions) => {
   const dailyVolume = options.createBalances();
+  const dailyFees = options.createBalances();
 
   const startTimestamp = options.startTimestamp;
   const endTimestamp = options.endTimestamp;
@@ -48,18 +49,25 @@ const fetch = async (options: FetchOptions) => {
   const poolDayDatas: PoolDayData[] = data?.poolDayDatas || [];
 
   let totalVolumeUSD = 0;
+  let totalFeesUSD = 0;
 
   for (const dayData of poolDayDatas) {
     totalVolumeUSD += parseFloat(dayData.volumeUSD || '0');
+    totalFeesUSD += parseFloat(dayData.feesUSD || '0');
   }
 
   dailyVolume.addUSDValue(totalVolumeUSD);
+  dailyFees.addUSDValue(totalFeesUSD);
 
-  return { dailyVolume };
+  return { dailyVolume, dailyFees, dailyRevenue: 0, dailySupplySideRevenue: dailyFees };
 };
 
 const methodology = {
-  Volume: 'Volume of all spot token swaps that go through the protocol.'
+  Volume: 'Volume of all spot token swaps that go through the protocol.',
+  Fees: 'Swap fees paid by users.',
+  UserFees: 'Swap fees paid by users.',
+  Revenue: 'No protocol revenue.',
+  SupplySideRevenue: 'All swap fees distributed to suppliers.',
 }
 
 const adapter: SimpleAdapter = {
@@ -67,7 +75,7 @@ const adapter: SimpleAdapter = {
   fetch,
   chains: [CHAIN.BASE],
   start: 1752451200,
-  methodology
+  methodology,
 };
 
 export default adapter;
